see if this cast fixes it?

+++ src/radar.c
@@ -513,7 +513,7 @@ int ret_radar_hk_fill(ret_radar_hk_t * h, ret_radar_hk_data_t *hk)
     rd += this_rd; 
   }
 
-  hk->board_temp = 0.0625 * ( resp[1] + (((int)resp[2]) <<8)); 
+  hk->board_temp = 0.0625 * ( (short) ( resp[1] + (((int)resp[2]) <<8))); 
   hk->air_temp = -256 + (1/32.) * ( resp[3] + (((int)resp[4]) <<8)); 
   hk->vin = 0.0515325 * ( resp[5] + (((int)resp[6]) <<8));
